vkCmdCopyImage: Allow copies between compatible formats.
The docs for the `MTLBlitCommandEncoder`'s `copyFromTexture:...` method
state that the source and destination textures must have the same
format.

The documentation lied.

Or rather, it wasn't a lie when it was written in the iOS 8 timeframe
for Metal 1.0. But, I suspect it was changed in Metal 1.1 when support
was added for texture views of differing formats. Like with texture
views, the two formats now need only be compatible. This is the same
behavior that Vulkan prescribes.