MVKDevice: Support the nullDescriptor feature.
My testing shows that Metal satisfies all the conditions for the
`nullDescriptor` feature from `VK_EXT_robustness2`--even for buffers. I
suspect that Metal does something similar to the `__PAGEZERO` segment in
64-bit host programs--that is, all device memory is allocated above 4 GB
in case of null buffer pointers.

Unfortunately, Apple's videos on the subject indicate that this won't
fly with their validation layers.