Add a function to query if a device supports presentation at all.
This function, `vkGetPhysicalDeviceMetalPresentationSupportEXT()`--which
I really need to add to the `VK_EXT_metal_surface` spec--will allow
clients to query if a physical device **can** be used to present
swapchain images at all, without needing to create a surface first. On
Mac, it returns the negated value of the `isHeadless` property; on iOS,
it always returns `true`.